Jinja2 optimization (#132)
* :hammer: further optimization. all jinja2 related code should go in jinja2 folder

* :tractor: all jinja2 extensions shall live in jinja2 folder.

* :micrscope: add jinja2 extensions test

* :green_heart: keep two stage lml loading. Althrough it is not yet understood why, it works.

* :bug: fix loading bug. otherwise, Error: jinja2

* :books: make notes on critical code lines
