+++ moban/jinja2/engine.py
@@ -1,9 +1,48 @@
 from jinja2 import Environment, FileSystemLoader
-from lml.plugin import PluginInfo
+from lml.loader import scan_plugins_regex
+from lml.plugin import PluginInfo, PluginManager
 
-import moban.utils as utils
 import moban.constants as constants
-from moban import plugins
+
+JINJA2_LIBRARIES = "^moban_jinja2_.+$"
+JINJA2_EXENSIONS = [
+    "moban.jinja2.filters.repr",
+    "moban.jinja2.filters.github",
+    "moban.jinja2.filters.text",
+    "moban.jinja2.tests.files",
+]
+
+
+class PluginMixin:
+    def get_all(self):
+        for name in self.registry.keys():
+            # only the first matching one is returned
+            the_filter = self.load_me_now(name)
+            yield (name, the_filter)
+
+
+class JinjaFilterManager(PluginManager, PluginMixin):
+    def __init__(self):
+        super(JinjaFilterManager, self).__init__(
+            constants.JINJA_FILTER_EXTENSION
+        )
+
+
+class JinjaTestManager(PluginManager, PluginMixin):
+    def __init__(self):
+        super(JinjaTestManager, self).__init__(constants.JINJA_TEST_EXTENSION)
+
+
+class JinjaGlobalsManager(PluginManager, PluginMixin):
+    def __init__(self):
+        super(JinjaGlobalsManager, self).__init__(
+            constants.JINJA_GLOBALS_EXTENSION
+        )
+
+
+FILTERS = JinjaFilterManager()
+TESTS = JinjaTestManager()
+GLOBALS = JinjaGlobalsManager()
 
 
 @PluginInfo(
@@ -18,6 +57,7 @@ class Engine(object):
 
         :param list temp_dirs: a list of template directories
         """
+        load_jinja2_extensions()
         self.template_dirs = template_dirs
         template_loader = FileSystemLoader(template_dirs)
         self.jj2_environment = Environment(
@@ -26,13 +66,13 @@ class Engine(object):
             trim_blocks=True,
             lstrip_blocks=True,
         )
-        for filter_name, filter_function in plugins.FILTERS.get_all():
+        for filter_name, filter_function in FILTERS.get_all():
             self.jj2_environment.filters[filter_name] = filter_function
 
-        for test_name, test_function in plugins.TESTS.get_all():
+        for test_name, test_function in TESTS.get_all():
             self.jj2_environment.tests[test_name] = test_function
 
-        for global_name, dict_obj in plugins.GLOBALS.get_all():
+        for global_name, dict_obj in GLOBALS.get_all():
             self.jj2_environment.globals[global_name] = dict_obj
 
     def get_template(self, template_file):
@@ -57,7 +97,7 @@ class Engine(object):
     def apply_template(self, template, data, output):
         """
         It is not expected this function to write content to file system.
-        Please just apply data inside the template and return utf-8 encoded
+        Please just apply data inside the template and return utf-8
         content.
 
         :param template: a jinja2 template from :class:`.get_template`
@@ -67,6 +107,9 @@ class Engine(object):
         template.globals["__target__"] = output
         template.globals["__template__"] = template.name
         rendered_content = template.render(**data)
-        rendered_content = utils.strip_off_trailing_new_lines(rendered_content)
-        rendered_content = rendered_content.encode("utf-8")
+        rendered_content = rendered_content
         return rendered_content
+
+
+def load_jinja2_extensions():
+    scan_plugins_regex(JINJA2_LIBRARIES, "moban", None, JINJA2_EXENSIONS)
