+++ moban/jinja2/filters/text.py
@@ -1,6 +1,6 @@
 import re
 
-from moban.extensions import JinjaFilter
+from moban.jinja2.extensions import JinjaFilter
 
 
 @JinjaFilter()
@@ -13,7 +13,7 @@ def split_length(input_line, length):
         yield line
     else:
         while True:
-            if " " in line[start : start + limit]:  # flake8: noqa
+            if " " in line[start : start + limit]:  # noqa
                 # go back and find a space
                 while limit > 0 and line[start + limit] != " ":
                     limit -= 1
@@ -25,7 +25,7 @@ def split_length(input_line, length):
                 ] != " ":
                     limit += 1
 
-            yield line[start : start + limit]  # flake8: noqa
+            yield line[start : start + limit]  # noqa
             start = start + limit + 1
             limit = length
             if len(line[start:]) < length or start + limit >= len(line):
