+++ moban/main.py
@@ -24,11 +24,11 @@ def main():
     """
     program entry point
     """
-    plugins.refresh_plugins()
     parser = create_parser()
     options = vars(parser.parse_args())
     HASH_STORE.IGNORE_CACHE_FILE = options[constants.LABEL_FORCE]
     moban_file = options[constants.LABEL_MOBANFILE]
+    load_engine_factory_and_engines()  # Error: jinja2 if removed
     if moban_file is None:
         moban_file = mobanfile.find_default_moban_file()
     if moban_file:
@@ -147,3 +147,7 @@ def handle_command_line(options):
         engine.number_of_templated_files()
     )
     return exit_code
+
+
+def load_engine_factory_and_engines():
+    plugins.make_sure_all_pkg_are_loaded()
